requirements.txt
google-generativeai
streamlit 
pandas 
scikit-learn 
joblib 
seaborn
pytest
python-dotenv